remove unnecessary {parse: true} statements
This removes `{parse: true}` from both our setting of a list's card
collection and of a card's items collection.

The former is not necessary because it would cause the Card#parse
method to be invoked, which removes `items` from the response. However,
only direct fetches of cards bring down an items attribute in the JSON.
Therefore, this will not be necessary when we have fetched a board or a
list.

The latter is not necessary because it would invoke the Item#parse
function, which we have not even written (and the Backbone default just
passes along the JSON response unchanged).