remove 'list-display' class (problem with ords)
We are using this 'list-display' class as the 'modelElement' for our
board show OrdView options to keep track of our lists to be sorted. The
class 'list-display' in that context refers to a list and everything
inside of it (heading, cards, form). However, we were also assigning the
heading to have the class 'list-display' as well. This was causing us to
have twice as many jQuery elements as we want when it comes time to run
our OrdView#saveOrds method, and having these extra jQuery elements was
basically causing the 'ord' attribute of the lists to be assigned quasi-
randomly, resulting in buggy, non-deterministic ordering of the lists.
This commit fixes that bug.